scraper/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
requests==2.31.0
beautifulsoup4==4.12.2
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
openai==0.28.1
whitenoise==6.5.0